reorganize sound code
