+++ components/sound/sound-visualizer.tsx
